+++ src/App.vue
@@ -1,5 +1,5 @@
 <template>
-  <div id="App-Wrapper" :class="[appWrapperClass, theme]">
+  <div id="App-Wrapper" :class="[appWrapperClass, theme]" :style="wrapperStyle">
     <div
       id="App-Container"
       class="app-container max-w-10/12 lg:max-w-screen-2xl px-3 lg:px-8"
@@ -17,9 +17,9 @@
         </router-view>
       </div>
     </div>
-    <Footer />
     <div id="loading-bar-wrapper" :class="loadingBarClass"></div>
   </div>
+  <Footer />
   <div class="App-Mobile-sidebar" v-if="isMobile">
     <div id="App-Mobile-Profile" class="App-Mobile-wrapper">
       <MobileMenu />
@@ -37,6 +37,7 @@ import {
   computed,
   defineComponent,
   onBeforeMount,
+  onMounted,
   onUnmounted,
   ref,
   watch
@@ -76,7 +77,10 @@ export default defineComponent({
       await appStore.fetchConfig().then(() => {
         metaStore.addScripts(appStore.themeConfig.site_meta.cdn.prismjs)
         // Change the favicon dynamically.
-        if (appStore.themeConfig.site_meta.favicon !== '') {
+        if (
+          appStore.themeConfig.site_meta.favicon &&
+          appStore.themeConfig.site_meta.favicon !== ''
+        ) {
           const link = document.querySelector("link[rel~='icon']")
           if (link)
             link.setAttribute('href', appStore.themeConfig.site_meta.favicon)
@@ -147,6 +151,20 @@ export default defineComponent({
       window.removeEventListener('resize', resizeHanler)
     })
 
+    const wrapperStyle = ref({ 'min-height': '100vh' })
+
+    onMounted(() => {
+      let wrapperHeight = screen.height
+      const footerEl = document.getElementById('footer')
+      const footerHeight = footerEl?.getBoundingClientRect().height
+      if (typeof footerHeight === 'number') {
+        wrapperHeight = wrapperHeight - footerHeight * 2
+      }
+      wrapperStyle.value = {
+        'min-height': wrapperHeight + 'px'
+      }
+    })
+
     /**
      * Watches the app loading status, adding the `nprogress-custom-parent`
      * class to the nprogress container when loading.
@@ -164,11 +182,20 @@ export default defineComponent({
       scripts: computed(() => metaStore.scripts),
       themeConfig: computed(() => appStore.themeConfig),
       headerImage: computed(() => {
-        return { backgroundImage: `url(${appStore.headerImage})` }
+        return {
+          backgroundImage: `url(${
+            appStore.headerImage
+          }), url(${require('@/assets/default-cover.jpg')})`,
+          opacity: appStore.headerImage !== '' ? 1 : 0
+        }
       }),
       headerBaseBackground: computed(() => {
-        return { background: appStore.themeConfig.theme.header_gradient_css }
+        return {
+          background: appStore.themeConfig.theme.header_gradient_css,
+          opacity: appStore.headerImage !== '' ? 0.81 : 0.91
+        }
       }),
+      wrapperStyle: computed(() => wrapperStyle.value),
       handleEscKey: appStore.handleEscKey,
       isMobile: computed(() => appStore.isMobile),
       appWrapperClass,
@@ -192,7 +219,7 @@ body {
   @apply relative min-w-full min-h-screen h-full;
   font-family: Rubik, Avenir, Helvetica, Arial, sans-serif;
   .app-wrapper {
-    @apply bg-ob-deep-900 min-w-full min-h-screen h-full;
+    @apply bg-ob-deep-900 min-w-full h-full pb-12;
     transition-property: transform, border-radius;
     transition-duration: 350ms;
     transition-timing-function: ease;
@@ -340,9 +367,12 @@ body {
 .app-banner-image {
   z-index: 1;
   background-size: cover;
+  opacity: 0;
+  transition: ease-in-out opacity 300ms;
 }
 
 .app-banner-screen {
+  transition: ease-in-out opacity 300ms;
   z-index: 2;
   opacity: 0.91;
 }