+++ src/components/ArticleCard/src/Article.vue
@@ -1,7 +1,8 @@
 <template>
   <li class="article">
     <div class="article-thumbnail">
-      <img v-if="post.cover" :src="post.cover" alt="article-thumbnail" />
+      <img v-if="post.cover" v-lazy="post.cover" alt="" />
+      <img v-else src="@/assets/default-cover.jpg" />
       <span class="thumbnail-screen" :style="gradientBackground" />
     </div>
     <div class="article-content">