+++ src/components/ArticleCard/src/HorizontalArticle.vue
@@ -1,12 +1,8 @@
 <template>
   <div class="feature-article">
     <div class="feature-thumbnail">
-      <img
-        v-if="post.cover"
-        class="ob-hz-thumbnail"
-        :src="post.cover"
-        alt="article-thumbnail"
-      />
+      <img v-if="post.cover" class="ob-hz-thumbnail" v-lazy="post.cover" />
+      <img v-else class="ob-hz-thumbnail" src="@/assets/default-cover.jpg" />
       <span class="thumbnail-screen" :style="bannerHoverGradient" />
     </div>
     <div class="feature-content">
@@ -49,7 +45,7 @@
         <div class="flex flex-row items-center">
           <img
             class="hover:opacity-50 cursor-pointer"
-            :src="post.author.avatar"
+            v-lazy="post.author.avatar"
             alt=""
             @click="handleAuthorClick(post.author.link)"
           />