+++ src/components/Sidebar/src/Profile.vue
@@ -11,18 +11,16 @@
       >
         <div class="flex flex-col justify-center items-center">
           <img
-            v-if="
-              themeConfig.site.avatar !== '' || themeConfig.site.logo !== ''
-            "
+            v-if="authorData.avatar !== ''"
             :class="avatarClass"
-            :src="themeConfig.site.avatar || themeConfig.site.logo"
+            :src="authorData.avatar"
             alt="avatar"
           />
           <ob-skeleton v-else width="7rem" height="7rem" circle />
 
           <h2 class="text-center pt-4 text-4xl font-semibold text-ob-bright">
-            <template v-if="themeConfig.site.author">
-              {{ themeConfig.site.author }}
+            <template v-if="authorData.name">
+              {{ authorData.name }}
             </template>
             <ob-skeleton v-else height="2.25rem" width="7rem" />
           </h2>
@@ -33,9 +31,9 @@
           />
 
           <p
-            v-if="themeConfig.site.description"
+            v-if="authorData.description"
             class="pt-6 px-10 w-full text-sm text-center"
-            v-html="themeConfig.site.description"
+            v-html="authorData.description"
           />
           <p
             v-else
@@ -45,75 +43,24 @@
           </p>
         </div>
         <div class="h-full w-full flex flex-col flex-1 justify-end items-end">
-          <ul
-            class="flex flex-row justify-evenly flex-wrap w-full py-4 px-2 text-center items-center"
-          >
-            <li class="diamond-clip-path diamond-icon" v-if="socials.github">
-              <a :href="socials.github" target="_blank" ref="github">
-                <svg-icon icon-class="github" class="fill-current" />
-              </a>
-            </li>
-            <li class="diamond-clip-path diamond-icon" v-if="socials.twitter">
-              <a :href="socials.twitter" target="_blank" ref="twitter">
-                <svg-icon icon-class="twitter" class="fill-current" />
-              </a>
-            </li>
-            <li
-              class="diamond-clip-path diamond-icon"
-              v-if="socials.stackoverflow"
-            >
-              <a
-                :href="socials.stackoverflow"
-                target="_blank"
-                ref="stackoverflow"
-              >
-                <svg-icon icon-class="stackoverflow" class="fill-current" />
-              </a>
-            </li>
-            <li class="diamond-clip-path diamond-icon" v-if="socials.wechat">
-              <a :href="socials.wechat" target="_blank" ref="wechat">
-                <svg-icon icon-class="wechat" class="fill-current" />
-              </a>
-            </li>
-            <li class="diamond-clip-path diamond-icon" v-if="socials.qq">
-              <a :href="socials.qq" target="_blank" ref="qq">
-                <svg-icon icon-class="qq" class="fill-current" />
-              </a>
-            </li>
-            <li class="diamond-clip-path diamond-icon" v-if="socials.weibo">
-              <a :href="socials.weibo" target="_blank" ref="weibo">
-                <svg-icon icon-class="weibo" class="fill-current" />
-              </a>
-            </li>
-            <li class="diamond-clip-path diamond-icon" v-if="socials.csdn">
-              <a :href="socials.csdn" target="_blank" ref="csdn">
-                <svg-icon icon-class="csdn" class="fill-current" />
-              </a>
-            </li>
-            <li class="diamond-clip-path diamond-icon" v-if="socials.zhihu">
-              <a :href="socials.zhihu" target="_blank" ref="zhifu">
-                <svg-icon icon-class="zhifu" class="fill-current" />
-              </a>
-            </li>
-            <li class="diamond-clip-path diamond-icon" v-if="socials.juejin">
-              <a :href="socials.juejin" target="_blank" ref="juejin">掘</a>
-            </li>
-          </ul>
+          <Social :socials="authorData.socials" />
           <ul class="grid grid-cols-4 pt-4 w-full px-2 text-lg">
             <li class="col-span-1 text-center">
-              <span class="text-ob-bright">{{ statistic.wordCount }}</span>
+              <span class="text-ob-bright">{{ authorData.word_count }}</span>
               <p class="text-base">{{ t('settings.words') }}</p>
             </li>
             <li class="col-span-1 text-center">
-              <span class="text-ob-bright">{{ statistic.posts }}</span>
+              <span class="text-ob-bright">{{
+                authorData.post_list.length
+              }}</span>
               <p class="text-base">{{ t('settings.articles') }}</p>
             </li>
             <li class="col-span-1 text-center">
-              <span class="text-ob-bright">{{ statistic.categories }}</span>
+              <span class="text-ob-bright">{{ authorData.categories }}</span>
               <p class="text-base">{{ t('settings.categories') }}</p>
             </li>
             <li class="col-span-1 text-center">
-              <span class="text-ob-bright">{{ statistic.tags }}</span>
+              <span class="text-ob-bright">{{ authorData.tags }}</span>
               <p class="text-base">{{ t('settings.tags') }}</p>
             </li>
           </ul>
@@ -124,19 +71,52 @@
 </template>
 
 <script lang="ts">
+import { AuthorPosts } from '@/models/Post.class'
 import { useAppStore } from '@/stores/app'
-import { computed, defineComponent, onMounted } from 'vue'
+import { useAuthorStore } from '@/stores/author'
+import { computed, defineComponent, onMounted, ref, toRefs, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
+import Social from '@/components/Social.vue'
 
 export default defineComponent({
   name: 'ObProfile',
-  setup() {
+  components: { Social },
+  props: {
+    author: {
+      type: String,
+      default: () => {
+        return ''
+      }
+    }
+  },
+  setup(props) {
     const appStore = useAppStore()
+    const authorStore = useAuthorStore()
     const { t } = useI18n()
+
+    const author = toRefs(props).author
+    const authorData = ref(new AuthorPosts())
 
     const fetchData = async () => {
       await appStore.fetchStat()
+      await fetchAuthor()
     }
+
+    const fetchAuthor = async () => {
+      if (author.value === '') return
+      await authorStore.fetchAuthorData(author.value).then((data) => {
+        authorData.value = data
+      })
+    }
+
+    watch(
+      () => props.author,
+      (newAuthor, oldAuthor) => {
+        if (newAuthor && newAuthor !== oldAuthor) {
+          fetchAuthor()
+        }
+      }
+    )
 
     onMounted(fetchData)
 
@@ -155,6 +135,7 @@ export default defineComponent({
         return appStore.themeConfig.socials
       }),
       statistic: computed(() => appStore.statistic),
+      authorData,
       t
     }
   }