+++ src/components/Sidebar/src/TagBox.vue
@@ -2,7 +2,7 @@
   <div class="sidebar-box">
     <SubTitle :title="'titles.tag_list'" icon="tag" />
     <TagList>
-      <template v-if="tags.length > 0">
+      <template v-if="tags && tags.length > 0">
         <TagItem
           v-for="tag in tags"
           :key="tag.slug"
@@ -23,8 +23,14 @@
           </span>
         </div>
       </template>
-      <template v-else>
+      <template v-else-if="tags">
         <ob-skeleton tag="li" :count="10" height="20px" width="3rem" />
+      </template>
+      <template v-else>
+        <div class="flex flex-row justify-center items-center">
+          <svg-icon class="stroke-ob-bright mr-2" icon-class="warning" />
+          {{ t('settings.empty-tag') }}
+        </div>
       </template>
     </TagList>
   </div>
@@ -51,7 +57,10 @@ export default defineComponent({
     onMounted(fetchData)
 
     return {
-      tags: computed(() => tagStore.tags),
+      tags: computed(() => {
+        if (tagStore.isLoaded && tagStore.tags.length === 0) return null
+        return tagStore.tags
+      }),
       t
     }
   }