+++ src/models/ThemeConfig.class.ts
@@ -7,6 +7,7 @@ interface ThemeRaw {
     theme: GeneralOptions
     site: StringConfig
     socials: StringConfig
+    custom_socials: GeneralOptions
     site_meta: GeneralOptions
     plugins: GeneralOptions
     version: string
@@ -36,6 +37,8 @@ export class ThemeConfig {
   site: Site = new Site()
   /** Socials config data */
   socials: Social = new Social()
+  /** Custom socials config data */
+  custom_socials: CustomSocials = new CustomSocials()
   /** Meta data for the site */
   site_meta: SiteMeta = new SiteMeta()
   /** Plugin configs */
@@ -56,6 +59,7 @@ export class ThemeConfig {
       this.theme = new Theme(rawConfig.theme)
       this.site = new Site(rawConfig.site)
       this.socials = new Social(rawConfig.socials)
+      this.custom_socials = new CustomSocials(rawConfig.custom_socials)
       this.plugins = new Plugins(rawConfig)
       this.site_meta = new SiteMeta(rawConfig.site_meta)
       this.version = rawConfig.version
@@ -293,6 +297,55 @@ export class Social {
           Object.assign(this, { [key]: raw[key] })
         }
       }
+    }
+  }
+}
+
+export class CustomSocial {
+  icon = {
+    iconfont: '',
+    img_link: ''
+  }
+  link = ''
+
+  constructor(raw?: GeneralOptions) {
+    if (raw) {
+      for (const key of Object.keys(this)) {
+        if (Object.prototype.hasOwnProperty.call(raw, key)) {
+          if (key === 'icon') {
+            if (
+              String(raw[key]).match(
+                /([a-zA-Z0-9\s_\\.\-():])+(.svg|.png|.jpg)$/g
+              )
+            ) {
+              console.log(raw[key])
+              Object.assign(this.icon, { img_link: raw[key] })
+            } else {
+              Object.assign(this.icon, { iconfont: raw[key] })
+            }
+          } else {
+            Object.assign(this, { [key]: raw[key] })
+          }
+        }
+      }
+    }
+  }
+}
+
+export class CustomSocials {
+  socials: CustomSocial[] = []
+
+  /**
+   * Model class for Social media links
+   *
+   * @param raw - Config data generated from Hexo
+   */
+  constructor(raw?: { [key: string]: any }) {
+    if (raw) {
+      Object.assign(
+        this.socials,
+        Object.keys(raw).map((key: string) => new CustomSocial(raw[key]))
+      )
     }
   }
 }